acionna0 can change n samples of adc conversion from PCY8575 by 6:n.. command

+++ components/modules/pcy8575/include/pcy8575.hpp
@@ -103,7 +103,7 @@ class pcy8575 {
 	void data_test(void);
 	uint8_t reset_reason(void);
 
-	int n_samples = N_SAMPLES;
+	uint16_t n_samples = N_SAMPLES;
 
 	uint16_t stream_array_raw[N_SAMPLES];
